Refine priority boosting throttling
* **Safeguard Throttling Counter:**
  Updated the throttling logic in `UpdatePriorityBoost` (in `scheduler.cpp`) to cast the result of `atomic_add(&sRescheduleCounter, 1)` to `uint32` before the modulo operation. This ensures consistent and safe behavior (wrapping modulo) even if the `int32` counter overflows to a negative value, avoiding potential undefined behavior with signed modulo operations in C++.

* **Ensure Atomic Consistency:**
  Verified that `fThreadCount` in `scheduler_cpu.cpp` is consistently accessed with atomic operations. Updated the reset logic in `RemoveCPU` to use `atomic_set(&fThreadCount, 0)` instead of a direct assignment, aligning with the atomic read/modify pattern used elsewhere in the file.

+++ src/system/kernel/scheduler/scheduler.cpp
@@ -84,7 +84,7 @@ UpdatePriorityBoost(CoreEntry* core, CPUEntry* cpu, ThreadData* running)
 
 	// Throttle priority boosting to improve scalability.
 	// We only run this O(N) operation every 50th reschedule.
-	if (atomic_add(&sRescheduleCounter, 1) % 50 != 0)
+	if ((uint32)atomic_add(&sRescheduleCounter, 1) % 50 != 0)
 		return;
 
 	/*
